链表的重复元素的处理
